requirements.txt
bnlp_toolkit==3.2.0
transformers
